Don't let reddit users PM to '/u/' prefaced users
This is probably not a big deal, but at the moment if you PM the bot to tip another user on reddit, and you include /u/ prefixed to the user's name, the bot will send coins to an account that can't exist. For instance: 
    !tip /u/jwinterm 1
would send coins to reddit:/u/jwinterm, but '/' is not a valid character for reddit usernames. This change presumes that the sender intended to send funds to jwinterm and strips of the '/u/' preceding the actual username.